+++ strongbox-storage/strongbox-storage-maven-layout-provider/src/main/java/org/carlspring/strongbox/services/impl/ArtifactSearchServiceImpl.java
@@ -1,7 +1,20 @@
 package org.carlspring.strongbox.services.impl;
 
+import java.io.IOException;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
+import org.apache.lucene.queryparser.classic.ParseException;
 import org.carlspring.strongbox.configuration.Configuration;
 import org.carlspring.strongbox.configuration.ConfigurationManager;
+import org.carlspring.strongbox.services.ArtifactEntryService;
 import org.carlspring.strongbox.services.ArtifactSearchService;
 import org.carlspring.strongbox.services.ConfigurationService;
 import org.carlspring.strongbox.storage.Storage;
@@ -12,15 +25,9 @@ import org.carlspring.strongbox.storage.indexing.SearchRequest;
 import org.carlspring.strongbox.storage.indexing.SearchResult;
 import org.carlspring.strongbox.storage.indexing.SearchResults;
 import org.carlspring.strongbox.storage.repository.Repository;
-
-import javax.inject.Inject;
-import java.io.IOException;
-import java.util.Collection;
-import java.util.Set;
-
-import org.apache.lucene.queryparser.classic.ParseException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 /**
@@ -31,102 +38,163 @@ public class ArtifactSearchServiceImpl
         implements ArtifactSearchService, ConfigurationService
 {
 
+    private static final String QUERY_PATTERN_DB = "([\\w]+)?=([^;]+);?";
+
     private static final Logger logger = LoggerFactory.getLogger(ArtifactSearchServiceImpl.class);
 
-    @Inject
+    @Autowired
     private RepositoryIndexManager repositoryIndexManager;
 
-    @Inject
+    @Autowired
     private ConfigurationManager configurationManager;
 
+    @Autowired
+    private ArtifactEntryService artifactEntryService;
 
     @Override
     public SearchResults search(SearchRequest searchRequest)
-            throws IOException, ParseException
+        throws IOException,
+        ParseException
     {
         SearchResults searchResults = new SearchResults();
 
-        final String repositoryId = searchRequest.getRepositoryId();
+        String storageId = searchRequest.getStorageId();
+        String repositoryId = searchRequest.getRepositoryId();
+        String query = searchRequest.getQuery();
+
+        Pattern pattern = Pattern.compile(QUERY_PATTERN_DB);
+        Matcher matcher = pattern.matcher(query);
+        // If query matches then we have a DataBase search request
+        if (matcher.matches())
+        {
+            Map<String, String> coordinates = new HashMap<>();
+            while (matcher.find())
+            {
+                coordinates.put(matcher.group(1), matcher.group(2));
+            }
+            searchResults.getResults()
+                         .addAll(dbSearch(coordinates, storageId, repositoryId));
+            return searchResults;
+        }
 
         final Collection<Storage> storages = getConfiguration().getStorages().values();
         if (repositoryId != null && !repositoryId.isEmpty())
         {
             logger.debug("Repository: {}", repositoryId);
-
-            final String storageId = searchRequest.getStorageId();
             if (storageId == null)
             {
-                for (Storage storage : storages)
-                {
-                    if (storage.containsRepository(repositoryId))
-                    {
-                        final String contextId = storage.getId() + ":" +
-                                                 repositoryId + ":" +
-                                                 IndexTypeEnum.LOCAL.getType();
-                        final Set<SearchResult> sr = repositoryIndexManager.getRepositoryIndexer(contextId)
-                                                                           .search(searchRequest.getQuery());
-
-                        if (sr != null && !sr.isEmpty())
-                        {
-                            searchResults.getResults().addAll(sr);
-                        }
-                    }
-                }
-
-                logger.debug("Results: {}", searchResults.getResults().size());
-
-                return searchResults;
+                searchResults.getResults().addAll(indexSearch(query, repositoryId, storages));
             }
             else
             {
-                String contextId = searchRequest.getStorageId() + ":" +
-                                   searchRequest.getRepositoryId() + ":" +
-                                   IndexTypeEnum.LOCAL.getType();
-
-                final Set<SearchResult> sr = repositoryIndexManager.getRepositoryIndexer(contextId)
-                                                                   .search(searchRequest.getQuery());
-
-                if (!sr.isEmpty())
-                {
-                    searchResults.getResults().addAll(sr);
-                }
-
-                logger.debug("Results: {}", searchResults.getResults().size());
-
-                return searchResults;
+                Storage storage = getConfiguration().getStorage(storageId);
+                searchResults.getResults().addAll(indexSearch(query, storage, repositoryId));
             }
         }
         else
         {
-            for (Storage storage : storages)
+            searchResults.getResults().addAll(indexSearch(query, storages));
+        }
+        logger.debug("Results: {}", searchResults.getResults().size());
+        return searchResults;
+    }
+
+    private List<SearchResult> dbSearch(Map<String, String> coordinates,
+                                        String storageId,
+                                        String repositoryId)
+    {
+        List<SearchResult> result = new LinkedList<>();
+
+        result.addAll(artifactEntryService.findByCoordinates(coordinates)
+                                          .stream()
+                                          .map(a -> new SearchResult(a.getStorageId(), a.getRepositoryId(),
+                                                  a.getArtifactCoordinates(), ""))
+                                          .collect(Collectors.toList()));
+        return result;
+    }
+
+    private List<SearchResult> indexSearch(String query,
+                                           Collection<Storage> storages)
+        throws ParseException,
+        IOException
+    {
+        List<SearchResult> result = new LinkedList<>();
+        for (Storage storage : storages)
+        {
+            for (Repository r : storage.getRepositories().values())
             {
-                for (Repository r : storage.getRepositories().values())
+                logger.debug("Repository: {}", r.getId());
+
+                final RepositoryIndexer repositoryIndex = repositoryIndexManager.getRepositoryIndexer(storage.getId()
+                        + ":" + r.getId());
+                if (repositoryIndex != null)
                 {
-                    logger.debug("Repository: {}", r.getId());
-
-                    String contextId = storage.getId() + ":" + r.getId() + ":" + IndexTypeEnum.LOCAL.getType();
-                    final RepositoryIndexer repositoryIndexer = repositoryIndexManager.getRepositoryIndexer(contextId);
-                    if (repositoryIndexer != null)
+                    final Set<SearchResult> sr = repositoryIndexManager.getRepositoryIndexer(storage.getId() + ":"
+                            + r.getId()).search(query);
+                    if (sr != null && !sr.isEmpty())
                     {
-                        final Set<SearchResult> sr = repositoryIndexer.search(searchRequest.getQuery());
-
-                        if (sr != null && !sr.isEmpty())
-                        {
-                            searchResults.getResults().addAll(sr);
-                        }
+                        result.addAll(sr);
                     }
                 }
             }
-
-            logger.debug("Results: {}", searchResults.getResults().size());
-
-            return searchResults;
         }
+        return result;
+    }
+
+    private List<SearchResult> indexSearch(String query,
+                                           Storage storage,
+                                           String repositoryId)
+        throws ParseException,
+        IOException
+    {
+        Repository repository = storage.getRepository(repositoryId);
+        String storageAndRepositoryId = getIndexId(storage, repository);
+        List<SearchResult> result = new LinkedList<>();
+        final Set<SearchResult> sr = repositoryIndexManager.getRepositoryIndexer(storageAndRepositoryId)
+                                                           .search(query);
+        if (!sr.isEmpty())
+        {
+            result.addAll(sr);
+        }
+        return result;
+    }
+
+    protected String getIndexId(Storage storage,
+                                Repository repository)
+    {
+        return storage.getId() + ":" + repository.getId() + ":"
+                + (repository.isProxyRepository() ? IndexTypeEnum.REMOTE.getType() : IndexTypeEnum.LOCAL.getType());
+    }
+
+    private List<SearchResult> indexSearch(String query,
+                                           String repositoryId,
+                                           Collection<Storage> storages)
+        throws ParseException,
+        IOException
+    {
+        List<SearchResult> result = new LinkedList<>();
+        for (Storage storage : storages)
+        {
+            if (storage.containsRepository(repositoryId))
+            {
+                final String storageAndRepositoryId = getIndexId(storage, storage.getRepository(repositoryId));
+                final Set<SearchResult> sr = repositoryIndexManager.getRepositoryIndexer(storageAndRepositoryId)
+                                                                   .search(query);
+
+                if (sr != null && !sr.isEmpty())
+                {
+                    result.addAll(sr);
+                }
+            }
+        }
+
+        return result;
     }
 
     @Override
     public boolean contains(SearchRequest searchRequest)
-            throws IOException, ParseException
+        throws IOException,
+        ParseException
     {
         return !search(searchRequest).getResults().isEmpty();
     }
