+++ strongbox-storage/strongbox-storage-maven-layout-provider/src/main/java/org/carlspring/strongbox/storage/indexing/SearchResults.java
@@ -1,13 +1,12 @@
 package org.carlspring.strongbox.storage.indexing;
 
+import java.util.LinkedHashSet;
+import java.util.Set;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
-import java.util.LinkedHashSet;
-import java.util.Set;
-
-import org.springframework.util.StringUtils;
 
 /**
  * @author mtodorov
@@ -43,15 +42,7 @@ public class SearchResults
         for (SearchResult artifact : results)
         {
             sb.append(artifact.getStorageId()).append(':').append(artifact.getRepositoryId()).append(' ');
-            sb.append(artifact.getGroupId()).append(':');
-            sb.append(artifact.getArtifactId()).append(':');
-            sb.append(artifact.getVersion()).append(':');
-            sb.append(artifact.getExtension());
-            if (!StringUtils.isEmpty(artifact.getClassifier()))
-            {
-                sb.append(':');
-                sb.append(artifact.getClassifier());
-            }
+            sb.append(artifact.getPath()).append(':');
             sb.append(' ');
             sb.append(artifact.getUrl());
             sb.append('\n');
